platform: use signed int for NMPlatformLnkBond.primary
On netlink API, the attribute is indeed u32. However, this is an ifindex
wihch in most other kernel APIs and in NetworkManager code is a signed
integer. Note that of course kernel would only ever assign numbers that
are valid ifindexes, thus in the suitable range.

Also, in _parse_lnk_bond() don't fallback to IFLA_BOND_ACTIVE_SLAVE.
The active slave is not the same as the primary.