fixup! libnm/client: don't reset properties when interface goes away
We want that when a D-Bus interface get removed (which in practice
means, the D-Bus object got removed altogether), that the properties are
frozen in time. Previously, they were reset to the default, which isn't
very useful.

... Except for the complex properties that refer other objects. We need
to break the reference cycles.

Previously, the patch only checked for the D-Bus types, and special
handled "o" and "ao" types. That seems wrong to me.

1) at the point where the check was performed, we didn't
  yet check whether the variant has the expected signature. If (due
  to a bug or an API break) the type changes, we still want to reset
  the property.
  For that, we would have to call_obj_handle_dbus_prop_changes() and
  not shortcut.

2) we have handlers, I think those handlers should be asked to handle
  the removal, and not the calling code decide that it's not necessary.

2b) nm_active_connection_get_specific_object_path() is D-Bus type "o",
  but the libnm property is a plain string. For this we should also
  freeze the value.

2b) is actually the only exception, where it really makes a difference.
Maybe that is a weak justification for this large patch, compared to the "simpler"
variant before. I think the "simpler" variant is however a bit "wrong", and
doing the wrong thing is usually simpler.