ovs-port: disambiguate "port" and "iface"
The conscious language fix in commit bcc958c41128 ('device: rename
{enslave,release}_slave() to {attach,detach}_port()'), introduced
the use of "port" as a generic term for a device that's subordinate
relationship with another.

In OpenVSwitch, though, "port" means a device that "interface" devices are
subordinate to. Trying to use the both meanings at the same time makes the
code supremely difficult to follow.

Let's make sure we're avoiding the generic term wherever possible and
clearly indicate we're using the OVS term.

While at it, let's remove the last vestiges of obsolete language and
also disambiguate "iface" (which means a different thing in NMPlatform
than it does in OVS).

https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/merge_requests/1371