n-dhcp4: close packet socket after timeout
In the DRAINING state we keep the packet socket alive until the next
I/O event, after the creation of the UDP socket. The goal is to avoid
missing packets when switching to the connected socket.

However, in normal conditions the next I/O event happens when the
address is renewed and therefore the packet socket stays around
unnecessarily for a long time.

Introduce a timeout of 10 seconds after which, without other events,
the packet socket gets closed.