glib-aux: use nm_assert() in nm_{ptr,}array_find_bsearch()
These checks don't seem very useful, to have them enabled
in production code.

What is actually the real danger of messing up with binary search,
is that the input array is not properly sorted. Asserting for that
would be way more useful, but also likely too expensive to be worth
it.

Checking that the input arguments are not NULL/zero, is not that useful,
because we "usually" won't make such mistakes.

While at it, declare each local variable on a separate line.